+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+* v0.0.7
+  - Add open source chef server to close #10 (using Ubuntu 12.04)
+  - Add chef box
+  - Add chef download
+  - Add `download:files` task
+  - Add `download.rb` file
+  - Add `locations/init_scripts` directory location for init scripts
+  - Add `locations/downloads` directory location for misc downloads
+  - Add `locations/repos` directory location for repos
+  - Moved repos to core.yml
+  - Add init script for chef server
+  - Add additional comments and terminal logging
+
 * v0.0.6
   - Add locale to dev init
   - Add apt-get update/autoclean/remove to init